% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bayesprofile_HMC.R
\name{Bayesprofile_HMC}
\alias{Bayesprofile_HMC}
\title{Bayesian Profiling using the Hamiltonian Monte Carlo Algorithm}
\usage{
Bayesprofile_HMC(dist, prmu0, prinf0, M, chains)
}
\arguments{
\item{prmu0}{mean of the prior distribution (vector)}

\item{prinf0}{the information matrix of prior distribution (matrix)}

\item{M}{MCMC iteration numbers (int)}

\item{chains}{burn-in iterations (int)}

\item{data}{data of both aggregate data and target list (dataframe/matrix)}
}
\value{
the fitted result as an instance of stanfit
}
\description{
This function estimates the probabilities of being included in the target list
conditional on unobserved group S using the observed indicator X using the Hamiltonian Monte Carlo Algorithm.
}
\examples{





}
