% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bayesprofile_MH.R
\name{Bayesprofile_MH}
\alias{Bayesprofile_MH}
\title{Bayesian Profiling using the Metropolis-Hastings Algorithm}
\usage{
Bayesprofile_MH(data, mu0, inf0, M, it0, flag_smooth = FALSE)
}
\arguments{
\item{data}{data of both aggregate data and target list (dataframe/matrix)}

\item{mu0}{mean of the prior distribution (vector)}

\item{inf0}{the information matrix of prior distribution (matrix)}

\item{M}{MCMC iteration numbers (int)}

\item{it0}{burn-in iterations (int)}

\item{flag_smooth}{Dirichlet smoothing or not (bool)}
}
\value{
A list consists of all interations of Markov Chain, the normalized markov chain,
probabilities of being included in the listin each chain, log-likelihood of each chain, and
whether this the morkov chain is rejected to move forward at each step.
}
\description{
This function estimates the probabilities of being included in the target list
conditional on unobserved group S using the observed indicator X using the Metropolis-Hastings Algorithm.
}
\examples{





}
