% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bayesprofiling-package.R
\docType{package}
\name{Bayesprofiling-package}
\alias{Bayesprofiling}
\alias{Bayesprofiling-package}
\title{Bayesprofiling: Bayesian Consumer Profiling: Estimate Consumer Characteristics from Aggregate Data}
\description{
This package is a realization of De Bruyn, Arnaud, and Thomas Otter (2022) \cite{Bayesian Consumer Profiling: How to Estimate Consumer Characteristics from Aggregate Data}, Journal of Marketing Research (2022), 59(4), 755–774. This package is corresponding to the idea of the paper, which is using the joint distribution of observed categories and the unobserved categories in the reference population and the distribution of observed categories in the target list to estimate the distribution of unobserved categories in the target list. We provide several methods, including the simple count method, Random Walk Markov Chain Monte Carlo (MCMC), Maximum Likelihood method, and we provide the example using Hamiltonian MC. The users could try different method (with different implicit assumptions) to compare the result.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hihowme/Bayesprofiling}
}

}
\keyword{internal}
