% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_dist.R
\name{data_dist}
\alias{data_dist}
\title{Distribution information of panel data}
\usage{
data_dist(data)
}
\arguments{
\item{data}{data of both aggregate data and target list (dataframe/matrix)}
}
\value{
a list of data distribution informations
}
\description{
Get the distribution information of data for further analysis
}
\examples{




}
