% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Election}
\alias{Election}
\title{Election}
\format{
A data frame with 23610 rows and 16 variables
}
\usage{
data(Election)
}
\description{
A data set containing the information of the votes of a targeting
group, and the population voting preference distribution of the zipcode
groups. There are 23610 zipcode groups (The zipcode group not in the targeting list
are not included). The variables are as follows:
}
\details{
\itemize{
  \item zipcode_group. The zipcode group. (1 to 23610)
  \item voter. The number of voters in each zipcode group.
  \item Other columns are the name of the president candidates.
}
}
\keyword{data}
