% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logmarglikGD.R
\name{logmarglikGD}
\alias{logmarglikGD}
\title{logmarglikGD}
\usage{
logmarglikGD(wmc, n_S, loglikmc, mu0, inf0, L)
}
\arguments{
\item{wmc}{data of both aggregate data and target list (dataframe/matrix)}

\item{n_S}{mean of the prior distribution (vector)}

\item{loglikmc}{the information matrix of prior distribution (matrix)}

\item{mu0}{MCMC iteration numbers (int)}

\item{inf0}{burn-in iterations (int)}

\item{L}{Dirichlet smoothing or not (bool)}
}
\value{
A list consists of all interations of Markov Chain, the normalized markov chain,
probabilities of being included in the listin each chain, log-likelihood of each chain, and
whether this the morkov chain is rejected to move forward at each step.
}
\description{
This function estimates the probabilities of being included in the target list
conditional on unobserved group S using the observed indicator X using the Metropolis-Hastings Algorithm.
}
\examples{

}
