% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Occupation}
\alias{Occupation}
\title{Occupation}
\format{
A data frame with 32557 rows and 15 variables
}
\usage{
data(Occupation)
}
\description{
A data set containing the information of the jobs of a targeting
group, and the population job distribution of the zipcode
groups. There are 32557 zipcodes.:
}
\details{
\itemize{
  \item zip_code. The zipcode group.
  \item visitors. The number of visitors in each zipcode group.
  \item Other columns are the job types.
}
}
\keyword{data}
