% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Occupation_population}
\alias{Occupation_population}
\title{Occupation_population}
\format{
A data frame with 32557 rows and 13 variables
}
\usage{
data(Occupation_population)
}
\description{
A data set containing the information of the career choices of the population
in each of the zipcode group. There are 32557 zipcodes.:
}
\details{
\itemize{
  \item zip_code. The zipcode group.
  \item Other columns are the job types.
}
}
\keyword{data}
