% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Occupation_target}
\alias{Occupation_target}
\title{Occupation_target}
\format{
A data frame with 2817 rows and 2 variables
}
\usage{
data(Occupation_target)
}
\description{
A data set containing the information of the number of the visitors
in each of the zipcode group. There are 2817 zipcodes.:
}
\details{
\itemize{
  \item zip_code. The zipcode group.
  \item visitors. The number of visitors in each zipcode group.
}
}
\keyword{data}
