% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Profile_MLE.R
\name{Profile_MLE}
\alias{Profile_MLE}
\title{Consumer Profiling using the Maximum Likelihood Method}
\usage{
Profile_MLE(data)
}
\arguments{
\item{data}{data of both aggregate data and target list (dataframe/matrix)}
}
\value{
the list of probability of being included in the target list and the optimization result given by nlm.
}
\description{
This function estimates the probabilities of being included in the target list
using the Maximum Likelihood Method.
}
\examples{






}
