% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_count.R
\name{simple_count}
\alias{simple_count}
\title{Profiling using the Simple count method}
\usage{
simple_count(dist)
}
\arguments{
\item{data}{data of both aggregate data and target list (dataframe/matrix)}
}
\value{
The probabilities of being included in the list
}
\description{
This function estimates the probabilities of being included in the target list
conditional on observed indicator X using the observed indicator X with the simple count method.
}
\examples{






}
