% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{VillagesSample}
\alias{VillagesSample}
\title{VillagesSample}
\format{
A data frame with 23610 rows and 2 variables
}
\usage{
data(VillagesSample)
}
\description{
A data set containing the information of the votes of a targeting
group. There are 23610 zipcode groups. The variables are as follows:
}
\details{
\itemize{
  \item zipcode_group. The zipcode group. (1 to 23610)
  \item voter. The number of voters in each zipcode group.
}
}
\keyword{data}
