% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{XS_pop}
\alias{XS_pop}
\title{Data}
\format{
An object of class \code{data.frame} with 36239 rows and 15 columns.
}
\usage{
XS_pop
}
\description{
Data
}
\references{
\url{data_blah.com}
}
\author{
Thomas Otter \email{blahblah@roxygen.org}
}
\keyword{data}
