% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bayesprofile_MH.R
\name{Bayesprofile_MH}
\alias{Bayesprofile_MH}
\title{Bayesian Profiling using the Metropolis-Hastings Algorithm}
\usage{
Bayesprofile_MH(dist, mu0, inf0, M, it0, flag_smooth = FALSE)
}
\arguments{
\item{dist}{a list containing the distribution information of both the aggregate data and target list (list)}

\item{mu0}{a vector specifying the mean of the prior distribution (vector)}

\item{inf0}{a matrix specifying the information matrix of prior distribution (matrix)}

\item{M}{an integer specifying MCMC iteration numbers (int)}

\item{it0}{an integer specifying burn-in iteration numbers (int)}

\item{flag_smooth}{an optional logical variable specifying Dirichlet smoothing or not (bool)}
}
\value{
\code{Bayesprofile_MH} returns a list containing the following fields:
\item{wmc}{the posterior markov chain}
\item{nwmc}{normalized markov chain}
\item{Smc}{probabilities of being included in the list in markov chain}
\item{loglikmc}{log-likelihood of markov chain}
\item{rejectmc}{the rejection record of the markov chain}
\item{dist_plot}{The plot of posterior distribution}
}
\description{
\code{Bayesprofile_MH} estimates the probabilities of being included in the target list
conditional on unobserved group S using the observed indicator X using the Metropolis-Hastings Algorithm.
}
\details{
Smc=Smc, loglikmc=loglikmc, rejectmc=rejectmc, dist_plot
}
\examples{

# data preparation
data(VillagesSample)
data(XS_pop)
Election = data_pre(VillagesSample,XS_pop,category = "zipcode_group",base= 'target')
# prior mu = 0, cov = 0.01, M = 4000, burn-in iterations = 200
Result_MH = Bayesprofile_MH(Election, 0, 0.01, 4000, 2000, FALSE)
# posterior distribution plot
Result_MH$dist_plot

}
\references{
De Bruyn, Arnaud, and Thomas Otter (2022)
\cite{Bayesian Consumer Profiling: How to Estimate Consumer Characteristics from Aggregate Data},
Journal of Marketing Research (2022), 59(4), 755–774
}
