% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_pre.R
\name{data_pre}
\alias{data_pre}
\title{Preprocessing of panel data}
\usage{
data_pre(target, population, category, base)
}
\arguments{
\item{population}{joint distribution of observed categories and unobserved categories (dataframe/matrix)}

\item{category}{name of the observed categories}

\item{base}{the base of the merge. if base = "population" then the intended list is the whole population and probability
0 would be assigned to those category that is in the list while not in the population,
if base = "target" then the target list is the range of consumer of interest.}
}
\value{
a panel data that contains information on both the target list and population distribution
}
\description{
Create a panel data that contains information on both the target list and population distribution
}
\examples{







}
