% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data1.R
\docType{data}
\name{VillagesSample}
\alias{VillagesSample}
\title{Data}
\format{
An object of class \code{data.frame} with 26310 rows and 2 columns.
}
\usage{
VillagesSample
}
\description{
Data
}
\references{
\url{data_blah.com}
}
\author{
Thomas Otter \email{blahblah@roxygen.org}
}
\keyword{data}
