% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{XS_pop}
\alias{XS_pop}
\title{XS_pop}
\format{
A data frame with 36239 rows and 15 variables
}
\usage{
data(XS_pop)
}
\description{
A data set containing the information of the voting preferences of the population
residents in each of the zipcode group. The 36239 zipcode groups are normalized to the
integer of 1 to 36239. The variables are as follows:
}
\details{
\itemize{
  \item zipcode_group. The zipcode group. (1 to 36239)
  \item Other columns are the name of the president candidates.
}
}
\keyword{data}
